🚚 Organise files
