+++ SmartCaptureDemo/API/ApiEndpoint.swift
