+++ SmartCaptureDemo/API/ApiManager.swift
