+++ SmartCaptureDemo/API/ApiManagerError.swift
