+++ SmartCaptureDemo/API/Models/ApplicantResponse.swift
@@ -1,0 +1,10 @@
+//
+//  ApplicantResponse.swift
+//  SmartCaptureDemo
+//
+//  Created by vincent.kearney on 13.03.24.
+//
+
+struct ApplicantResponse: Decodable {
+    let id: String
+}
