+++ SmartCaptureDemo/API/Models/SDKTokenResponse.swift
@@ -1,0 +1,10 @@
+//
+//  SDKTokenResponse.swift
+//  SmartCaptureDemo
+//
+//  Created by vincent.kearney on 13.03.24.
+//
+
+struct SDKTokenResponse: Decodable {
+    let token: String
+}
