+++ SmartCaptureDemo/API/Models/WorkFlowRunResponse.swift
@@ -1,0 +1,10 @@
+//
+//  WorkFlowRunResponse.swift
+//  SmartCaptureDemo
+//
+//  Created by vincent.kearney on 13.03.24.
+//
+
+struct WorkFlowRunResponse: Decodable {
+    let id: String
+}
