+++ SmartCaptureDemo/EnvironmentVariables/EnvironmentVars.swift
