+++ SmartCaptureDemo/Extensions/HTTPURLResponse+Ext.swift
