+++ SmartCaptureDemo/Extensions/WebView+Ext.swift
