+++ SmartCaptureDemo/Resources/Assets.xcassets/AppIcon.appiconset/1024.png
