+++ SmartCaptureDemo/Resources/Assets.xcassets/Contents.json
