+++ SmartCaptureDemo/Resources/Assets.xcassets/primaryDarkColor.colorset/Contents.json
