+++ SmartCaptureDemo/Resources/Assets.xcassets/primaryLightColor.colorset/Contents.json
