+++ SmartCaptureDemo/Resources/index.html
