+++ SmartCaptureDemo/Resources/webview_script.js
