+++ SmartCaptureDemo/UI/ConnectionView.swift
